Tweak docs, minor re-organization.
