+++ docs/spelling_wordlist.txt
@@ -65,6 +65,7 @@ protobuf
 proxied
 QueryOptions
 reimplemented
+Redis
 RequestHandler
 runtime
 schemas
