+++ tests/conftest.py
@@ -23,9 +23,9 @@ import os
 from unittest import mock
 
 from google.cloud import environment_vars
-from google.cloud.ndb import _cache
 from google.cloud.ndb import context as context_module
 from google.cloud.ndb import _eventloop
+from google.cloud.ndb import global_cache as global_cache_module
 from google.cloud.ndb import model
 
 import pytest
@@ -52,7 +52,7 @@ def reset_state(environ):
     yield
     model.Property._FIND_METHODS_CACHE.clear()
     model.Model._kind_map.clear()
-    _cache._InProcessGlobalCache.cache.clear()
+    global_cache_module._InProcessGlobalCache.cache.clear()
 
 
 @pytest.fixture
@@ -103,8 +103,8 @@ def in_context(context):
 def global_cache(context):
     assert not context_module._state.context
 
-    global_cache = _cache._InProcessGlobalCache()
-    with context.new(global_cache=global_cache).use():
-        yield global_cache
+    cache = global_cache_module._InProcessGlobalCache()
+    with context.new(global_cache=cache).use():
+        yield cache
 
     assert not context_module._state.context
